requirements.txt
playwright==1.52.0
pytest==8.3.5
pytest-playwright==0.7.0
python-dotenv==1.0.1
allure-pytest==2.14.0
requests==2.32.
pytest-xdist==3.8.0